+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 * Updated node to 6.2.0
 * Updated npm to 3.9.2
 * Removed typings from git
+* Updated typings
+* Updated all node packages
+* Fixed problem with nonexistent static directory
+* Fixed webpack validator custom schema config
 
 ### 1.5.0 (13.05.2016)
 * Updated Spring Boot to 1.3.4